se mueven todas las funciones existentes del routes a controller
